requirements.txt
torch==2.6.0 #+cu124
#h5py==3.8.0
opencv-python==4.7.0.72 
timm==0.9.2
albumentations==1.3.1
tensorboard==2.11.2 
gdown==4.7.1
numpy==1.24.4
matplotlib==3.7
segmentation-models-pytorch
Flask==2.2.5
Werkzeug==2.2.3
requests
shapely=2.1.2
flask_cors=6.0.2
rasterio=1.4.4
pyproj=3.7.2
cv2geojson=0.0.9